ivms: remove unknown XX country code
As detailed on the Wikipedia page that was previously
linked to, this convention is only used by the World
Intellectual Property Organization (WIPO), which is
irrelevant in the Travel Rule context.

We also take the opportunity to remove the minimally
useful ISO standard dependency, and use the already
available country list instead.